fixed error in finance calculator - cast input to float instead of int

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 monthly_income = input("Enter your monthly income:")
 monthly_expenses = input("Enter your total monthly expenses:")
 
-monthly_savings = int(monthly_income) - int(monthly_expenses)
+monthly_savings = float(monthly_income) - float(monthly_expenses)
 
 annual_interest_rate = 0.05
